feat : 이슈 클릭 핸들러 수정
