+++ src/components/IssueList.tsx
@@ -1,4 +1,5 @@
 import { Fragment } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 
 import { PER_PAGE } from '@/api/issue';
@@ -13,24 +14,29 @@ function isAdCell(index: number) {
 }
 
 export function IssueList() {
+  const navigate = useNavigate();
   const [issues, isLoading, hasNextPage, fetchNextPage] = useFetchIssues(PER_PAGE);
   const ref = useInfiniteScroll(async (entry, observer) => {
     observer.unobserve(entry.target);
     if (!isLoading && hasNextPage) fetchNextPage();
   });
 
+  const handleClickIssue = (number: number) => navigate(`${number}`);
+
   return (
     <div>
       {issues.map((issue, index) => (
         <Fragment key={issue.id}>
           {isAdCell(index) && <Advertise />}
-          <IssueInfo
-            issueNumber={issue.number}
-            title={issue.title}
-            author={issue.user?.login}
-            created_at={issue.created_at}
-            comments={issue.comments}
-          />
+          <Box onClick={() => handleClickIssue(issue.number)}>
+            <IssueInfo
+              issueNumber={issue.number}
+              title={issue.title}
+              author={issue.user?.login}
+              created_at={issue.created_at}
+              comments={issue.comments}
+            />
+          </Box>
         </Fragment>
       ))}
       <Target ref={ref} />
@@ -38,6 +44,13 @@ export function IssueList() {
     </div>
   );
 }
+
+const Box = styled.div`
+  padding: 20px;
+  box-shadow: 0px 0px 5px 0px #bcbcbc;
+  margin-bottom: 20px;
+  cursor: pointer;
+`;
 
 const Target = styled.div`
   height: 1px;
