+++ src/context/Context.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+import { IssueData } from '@/api/issue';
+
+export const IssueContext = createContext<IssueData[]>([]);
